Refactored Code and Fixed MakeList For Unix
