+++ include/VectorUtil.h
