+++ src/mex_files/trainLinFSA.cpp
@@ -1,7 +1,7 @@
 #include <math.h>
 #include <stdlib.h>
 #include "mex.h"
-#include "fsa.h"
+#include "Fsa.h"
 using namespace std;
 
 void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[]) {
